requirements.txt
API==0.0.7
Flask==2.1.1
Flask_Bcrypt==1.0.1
Flask_Cors==4.0.0
Flask_RESTful==0.3.10
PyJWT==2.4.0
PyJWT==2.8.0
PyMySQL==1.1.0
